modification(tests): update conversion testset
updating the testset for the new conversion done in the record service
instead in the  metadata schema class
